country filter fix some tests filter chain example

+++ src/main/java/io/github/jklingsporn/vertx/push/filters/LastSessionFilter.java
@@ -6,21 +6,21 @@ import io.github.jklingsporn.vertx.push.filters.relations.LessRelation;
 /**
  * Created by jensklingsporn on 04.01.17.
  */
-public interface LastSessionFilter extends GreaterRelation<Integer>,LessRelation<Integer> {
+public interface LastSessionFilter extends GreaterRelation<Float>,LessRelation<Float> {
 
     default String getContext(){
         return "last_session";
     }
 
     @Override
-    default Filter greater(Integer value) {
+    default Filter greater(Float value) {
         FilterImpl filter = new FilterImpl(getContext());
         filter.content().put("relation", ">").put("hours_ago", value.toString());
         return filter;
     }
 
     @Override
-    default Filter less(Integer value) {
+    default Filter less(Float value) {
         FilterImpl filter = new FilterImpl(getContext());
         filter.content().put("relation", "<").put("hours_ago", value.toString());
         return filter;
